mr/entities search and draw path (#20)
* Added entities_search

* Added draw path

* Move draw_path to another file

* Center window on the player

* Light bug fix

The player doesn't need to move before the light is shown

* Fix animation function name

Would lead to linking errors when used

* Player path integration with animation system

Lots of previous code had to be rewritten (sorry Mariana, this code will
forever be in our hearts) for this integration with this architecture

* Player movement only to lit up zones

* Auditory feedback for invalid movement

* Fixed spaces

* Trailing whitespace fixes

* Port of the A* algorithm to the animation system

Way less changes needed this time

* A* testing results in segfault on node freeing

A test function for A* shows flaws in this algorithm's implementation.
It fails while freeing nodes. Gcc was correct about the warnings.

* Solved segfault in A*, changed to Manhattan distance

* Fixed memory leak

The closest entity to the player still isn't moving. A bug in the A*
algorithm implementation is possible

---------

Co-authored-by: voidbert <50591320+voidbert@users.noreply.github.com>